seed-data/requirements.txt
bson
pymongo
python-dotenv
